merge branch 'development' to tests

+++ middlewares/jwt.go
@@ -18,11 +18,11 @@ func GenerateToken(userId int) (string, error) {
 	return token.SignedString([]byte(config.Env("JWT_SECRET")))
 }
 
-func ExtractToken(e echo.Context) int {
+func ExtractToken(e echo.Context) float64 {
 	user := e.Get("user").(*jwt.Token)
 	if user.Valid {
 		claims := user.Claims.(jwt.MapClaims)
-		return claims["userId"].(int)
+		return claims["userId"].(float64)
 	}
 	return 0
 }
